Rename components/cards/CardRow.vue to components/index/_shared/CardRow.vue
