+++ components/index/_shared/CardsLazyRow.vue
@@ -20,7 +20,7 @@
 <script lang="ts">
 import Vue from 'vue'
 
-import CardRow from '@/components/cards/CardRow.vue'
+import CardRow from '@/components/index/_shared/CardRow.vue'
 
 type Data = {
   actives: boolean[]
